Common: Consolidate agent_events and agent_event_serializers
The majority of the contents of these packages have been moved to
monkeyevents. Both of these packages now contain only code for
registering agent events and serializers. This code is cohesive enough
to constitute a single module instead of two.
