+++ monkey/infection_monkey/island_api_client/http_island_api_client.py
@@ -22,7 +22,7 @@ from requests import Response
 
 from common import AgentHeartbeat, AgentRegistrationData, AgentSignals
 from common.agent_configuration import AgentConfiguration
-from common.agent_event_serializers import AgentEventSerializerRegistry
+from common.agent_events import AgentEventSerializerRegistry
 from common.agent_plugins import AgentPlugin
 from common.common_consts.timeouts import SHORT_REQUEST_TIMEOUT
 from common.common_consts.token_keys import ACCESS_TOKEN_KEY_NAME, TOKEN_TTL_KEY_NAME
