+++ monkey/tests/unit_tests/common/agent_events/test_agent_event_serializer_registry.py
@@ -4,7 +4,7 @@ import pytest
 from monkeyevents import AbstractAgentEvent, IAgentEventSerializer
 from pydantic import Field
 
-from common.agent_event_serializers import AgentEventSerializerRegistry
+from common.agent_events import AgentEventSerializerRegistry
 
 
 class SomeEvent(AbstractAgentEvent):
