+++ monkey/tests/unit_tests/infection_monkey/island_api_client/test_http_island_api_client.py
@@ -22,7 +22,7 @@ from tests.data_for_tests.propagation_credentials import CREDENTIALS_DICTS
 
 from common import AgentRegistrationData, AgentSignals
 from common.agent_configuration import AgentConfiguration
-from common.agent_event_serializers import AgentEventSerializerRegistry
+from common.agent_events import AgentEventSerializerRegistry
 from common.common_consts.token_keys import ACCESS_TOKEN_KEY_NAME, TOKEN_TTL_KEY_NAME
 from infection_monkey.island_api_client import (
     HTTPIslandAPIClient,
