+++ monkey/tests/unit_tests/monkey_island/cc/repositories/test_mongo_agent_event_repository.py
@@ -10,7 +10,7 @@ from pydantic import Field
 from pymongo import MongoClient
 from tests.unit_tests.monkey_island.cc.repositories.mongo import get_all_collections_in_mongo
 
-from common.agent_event_serializers import AgentEventSerializerRegistry
+from common.agent_events import AgentEventSerializerRegistry
 from monkey_island.cc.repositories import (
     IAgentEventRepository,
     MongoAgentEventRepository,
